b3tracker/requirements.txt
asgiref==3.8.1
Django==5.0.6
sqlparse==0.5.0
typing_extensions==4.12.2
python-dotenv==1.0.1
celery==5.4.0
djangorestframework==3.15.2
markdown==3.6
django-filter==24.2
redis==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-redis==5.4.0
requests==2.32.3